make the last parameter an int and see if it worked
git-svn-id: file:///home/oestrich/tmp/img/trunk@1021 ea04b438-0a61-0410-93d0-9bc465a3bab0

+++ modules/albums/class.php
@@ -237,8 +237,8 @@ class Albums implements Module{
 					$this->lncln->page = 1;
 				}
 				else{
-					if(is_numeric(end($this->lncln->params))){
-						$this->lncln->page = end($this->lncln->params);	
+					if(is_numeric((int)end($this->lncln->params))){
+						$this->lncln->page = (int)end($this->lncln->params);	
 					}
 					else{
 						$this->lncln->page = 1;
